+++ source/EngineImpl/DescriptionConverter.cpp
@@ -1,5 +1,6 @@
 #include "DescriptionConverter.h"
 
+#include <cmath>
 #include <algorithm>
 #include <boost/range/adaptor/map.hpp>
 
@@ -509,7 +510,7 @@ namespace
 {
     void checkAndCorrectInvalidEnergy(float& energy)
     {
-        if (isnan(energy) || energy < 0 || energy > 1e12) {
+        if (std::isnan(energy) || energy < 0 || energy > 1e12) {
             energy = 0;
         }
     }